+++ packages/internal-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @defuse-protocol/internal-utils
 
+## 0.19.1
+
+### Patch Changes
+
+- Updated dependencies [2d38a96]
+  - @defuse-protocol/contract-types@0.3.1
+
 ## 0.19.0
 
 ### Minor Changes